feat: add keras tests
