+++ finetuner/tailor/keras/projection_head.py
@@ -11,24 +11,24 @@ class ProjectionHead(tf.keras.layers.Layer):
 
     def __init__(self, in_features: int, output_dim: int = 128, num_layers: int = 3):
         super().__init__()
-        self.head_layers = []
+        self.layers = []
         is_last_layer = False
         for idx in range(num_layers):
             if idx == num_layers - 1:
                 is_last_layer = True
             if not is_last_layer:
-                self.head_layers.append(
+                self.layers.append(
                     tf.keras.layers.Dense(
                         units=in_features,
                         bias_initializer='zeros',
                     )
                 )
-                self.head_layers.append(
+                self.layers.append(
                     tf.keras.layers.BatchNormalization(epsilon=self.EPSILON)
                 )
-                self.head_layers.append(tf.keras.layers.ReLU())
+                self.layers.append(tf.keras.layers.ReLU())
             else:
-                self.head_layers.append(
+                self.layers.append(
                     tf.keras.layers.Dense(
                         units=output_dim,
                         bias_initializer='zeros',
@@ -36,6 +36,6 @@ class ProjectionHead(tf.keras.layers.Layer):
                 )
 
     def call(self, x):
-        for layer in self.head_layers:
+        for layer in self.layers:
             x = layer(x)
         return x
